README; simplifies install instructions
If this commit's added to the README; the docs and readme instructions will
be identical.

This version of the instructions is more streamlined.
Although there's less hand holding, the beginning of the instructions
links to the "Troubleshooting" section on the docs' page.

That troubleshooting section has explicit directions for what to do at each
step. So I think this PR lets us have the best of the both worlds--streamlined
README for experienced users, and a guided docs site for those who need a bit
more help.